Rename database/sql
