+++ libs/database/sql/parse_test.go
@@ -1,4 +1,4 @@
-package sqltraced
+package sql
 
 import (
 	"reflect"
